Create a module for total comments and get comments
