+++ src/index.js
@@ -1,33 +1,9 @@
 import './style.css';
 import Logo from './images/google_books.png';
 import totalBooksItems from './modules/totalItems.js';
+import {getBookData, createComment} from './modules/totalComments';
 
 const bookList = document.getElementById('book-list');
-
-const createComment = async (bookId, user, comment) => {
-  const response = await fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/gcxOnR7Ou6sAxWdfnAQw/comments', {
-    method: 'POST',
-    body: JSON.stringify({
-      item_id: bookId,
-      username: user,
-      comment,
-    }),
-    headers: {
-      'Content-type': 'application/json; charset=UTF-8',
-    },
-  });
-  const jsonResponse = await response.json();
-  return jsonResponse;
-};
-
-const getBookData = async (bookId) => {
-  const res = await fetch(`https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/gcxOnR7Ou6sAxWdfnAQw/comments?item_id=${bookId}`);
-  const data = await res.json();
-  if (!data.error) {
-    return data;
-  }
-  return [];
-};
 
 const popUp = async (item) => {
   item.comments = await (getBookData(item.id));
