+++ src/modules/totalComments.js
@@ -1,0 +1,29 @@
+const createComment = async (bookId, user, comment) => {
+  const response = await fetch('https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/gcxOnR7Ou6sAxWdfnAQw/comments', {
+    method: 'POST',
+    body: JSON.stringify({
+      item_id: bookId,
+      username: user,
+      comment,
+    }),
+    headers: {
+      'Content-type': 'application/json; charset=UTF-8',
+    },
+  });
+  const jsonResponse = await response.json();
+  return jsonResponse;
+};
+
+const getBookData = async (bookId) => {
+  const res = await fetch(`https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/gcxOnR7Ou6sAxWdfnAQw/comments?item_id=${bookId}`);
+  const data = await res.json();
+  if (!data.error) {
+    return data;
+  }
+  return [];
+};
+
+export {
+  createComment, 
+  getBookData
+};
